Derive from TestCase. Rename so test classes/files match
