+++ tests/neuro_san/internals/run_context/langchain/core/test_base_tool_factory.py
@@ -18,6 +18,8 @@ from typing import Any
 from typing import Dict
 
 from copy import deepcopy
+
+from unittest import TestCase
 from unittest.mock import AsyncMock
 from unittest.mock import MagicMock
 
@@ -27,7 +29,7 @@ from neuro_san.internals.run_context.langchain.core.base_tool_factory import Bas
 from neuro_san.internals.utils.external_agent_parsing import ExternalAgentParsing
 
 
-class TestBaseToolFactory:
+class TestBaseToolFactory(TestCase):
     """
     Test cases for BaseToolFactory.
