+++ tests/neuro_san/internals/run_context/langchain/core/test_langchain_openai_function_tool.py
@@ -15,6 +15,7 @@
 #
 # END COPYRIGHT
 
+from unittest import TestCase
 from unittest.mock import AsyncMock
 from unittest.mock import MagicMock
 
@@ -29,7 +30,7 @@ from neuro_san.internals.run_context.langchain.core.tool_spec_error import ToolS
 from neuro_san.message.types.agent_tool_result_message import AgentToolResultMessage
 
 
-class TestLangChainOpenAIFunctionTool:
+class TestLangChainOpenAIFunctionTool(TestCase):
     """
     Test cases for what _arun() hands back to langchain as the tool output.
