+++ tests/neuro_san/internals/run_context/langchain/core/test_run_context_runnable.py
@@ -15,6 +15,7 @@
 #
 # END COPYRIGHT
 
+from unittest import TestCase
 from unittest.mock import AsyncMock
 from unittest.mock import MagicMock
 
@@ -35,7 +36,7 @@ from neuro_san.message.types.agent_framework_message import AgentFrameworkMessag
 from tests.neuro_san.message.content_fixtures import ContentFixtures
 
 
-class TestRunContextRunnable:
+class TestRunContextRunnable(TestCase):
     """
     Tests for RunContextRunnable: the capture-side projection of list-form
     (block) content in parse_chain_result, and the surfacing of
