+++ tests/neuro_san/internals/run_context/langchain/journaling/test_journaling_callback_handler.py
@@ -16,6 +16,8 @@
 # END COPYRIGHT
 
 from typing import Tuple
+
+from unittest import TestCase
 from unittest.mock import AsyncMock
 from unittest.mock import MagicMock
 from uuid import uuid4
@@ -32,7 +34,7 @@ from neuro_san.message.types.agent_message import AgentMessage
 from tests.neuro_san.message.content_fixtures import ContentFixtures
 
 
-class TestJournalingCallbackHandler:
+class TestJournalingCallbackHandler(TestCase):
     """
     Tests for JournalingCallbackHandler.
