+++ tests/neuro_san/internals/run_context/langchain/journaling/test_journaling_exclusive_agent_attribution.py
@@ -15,6 +15,8 @@
 # END COPYRIGHT
 
 from typing import Tuple
+
+from unittest import TestCase
 from unittest.mock import AsyncMock
 from unittest.mock import MagicMock
 from uuid import uuid4
@@ -27,7 +29,7 @@ from langchain_core.outputs import LLMResult
 from neuro_san.internals.run_context.langchain.journaling.journaling_callback_handler import JournalingCallbackHandler
 
 
-class TestExclusiveAgentAttribution:
+class TestJournalingExclusiveAgentAttribution(TestCase):
     """An inherited handler must ignore events owned by a descendant agent."""
 
     @staticmethod
