+++ tests/neuro_san/internals/run_context/langchain/llms/test_openai_llm_policy.py
@@ -24,6 +24,8 @@ from typing import Dict
 from typing import Iterator
 from typing import List
 
+from unittest import TestCase
+
 import pytest
 
 from langchain_core.messages import HumanMessage
@@ -33,7 +35,7 @@ from openai.resources.responses import AsyncResponses
 from neuro_san.internals.run_context.langchain.llms.openai_llm_policy import OpenAILlmPolicy
 
 
-class TestOpenAILlmPolicy:
+class TestOpenAILlmPolicy(TestCase):
     """
     Test cases for OpenAILlmPolicy.create_llm().
