+++ tests/neuro_san/internals/run_context/langchain/mcp/test_langchain_mcp_adapter.py
@@ -15,6 +15,7 @@
 #
 # END COPYRIGHT
 
+from unittest import TestCase
 from unittest.mock import AsyncMock
 from unittest.mock import MagicMock
 from unittest.mock import patch
@@ -25,7 +26,7 @@ from langchain_core.tools import StructuredTool
 from neuro_san.internals.run_context.langchain.mcp.langchain_mcp_adapter import LangChainMcpAdapter
 
 
-class TestLangChainMcpAdapter:
+class TestLangChainMcpAdapter(TestCase):
     """Test suite for LangChainMcpAdapter class"""
 
     @pytest.fixture
