+++ tests/neuro_san/internals/run_context/langchain/toolbox/test_toolbox_factory.py
@@ -19,6 +19,7 @@ import logging
 
 from unittest.mock import patch, MagicMock
 
+from unittest import TestCase
 import pytest
 
 from langchain_core.tools.base import BaseTool
@@ -35,7 +36,7 @@ VALIDATIOR_PATH = (
 )
 
 
-class TestToolboxFactory:
+class TestToolboxFactory(TestCase):
     """Simplified test suite for ToolboxFactory."""
 
     @pytest.fixture
